model/h5_deepx/requirements.txt
h5py
numpy
pyyaml
os
sys
